pyproject.toml
[tool.poetry]
name = "qleoz12 build-high-performance-web-apps-with-fastapi-by-malhar-lathkar main ch7-apirouter"
version = "0.1.0"
description = ""
authors = ["qleoz12 <kintaroloop@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.111.1"}
sqlalchemy = "^2.0.31"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
